requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Authlib==1.3.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
openai==1.23.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pip==23.3.1
protobuf==4.25.3
pyarrow==16.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
st-weaviate-connection==0.0.1
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.0
weaviate-client==3.26.2
wheel==0.41.2
